Only build on GCC7 on Linux
GCC4 doesn't support C++14, which versions >= 0.69 require.

*Note*: I was stuck for a *long* time trying to get this to work ... it turned
out that the simultaneous specification of `skip: true # [win or not py36]`
plus `host: - python 3.6` led to a straaange problem where one rerender would
leave the recipe unchanged, but a second one would make *all* of the build
variants disappear -- ?!?!? I spent a while thinking that the problem was due
to the new `cxx_compiler` selector I was trying to introduce, but the problem
turns out to have been totally unrelated.