Fix location of data files on Windows
We can take advantage of existing "relocation" support in poppler/GlobalParams.cc --
it's not convenient to use Conda's prefix rewriting because the prefix path has
to be in "mixed" form (`C:/foo/bar`), I think, and we'd need to add some custom code
to deal with that. On the other hand, we can essentially use the relocation code,
except for a patch to find the data files in $PREFIX/share/poppler while the DLL lives
in $PREFIX/Library/bin/poppler.dll. This, in turn, happens because the poppler-data
package is noarch and so has a common file layout on Unixes and Windows.