Attempt to fix TIFF I/O on Windows
From looking into #111, it seems that we were crashing when the
TiffWriter was used because the TIFF file handle was being closed
twice, which is a no-no on Windows. The TIFF writer needs to call
TIFFClose() to clean up resources, and that automatically closes
the TIFF file handle, but the semantics of the poppler "writer"
classes are such that it's the caller's responsibility to close
their file handles. Fortunately, it looks like things work if
we dup() the file handle.

Closes #111.