Update build flags and add Qt in the main build
... but don't depend on it.

The new release requires us to explicitly disable GPGME and QT6 in
the build configuration.

We are also running into an issue related to the multi-output package
build. The `install.sh` script tries to just do a `ninja install` to
install the built files for the two subpackages (`poppler` and
`poppler-qt`). However, the build system recreates and renames the
environments between `build.sh` and `install.sh`, and `ninja` notices
that and wants to rerun CMake. With the new behavior causing the issues
above, the CMake configuration fails for the `poppler` package because
Qt5 isn't available.

The easiest thing to try is to include `qt-main` as a host dependency
for the `poppler` package, making sure that it doesn't become a runtime
dependency. For reasons I don't understand, it seems that when we add an
`ignore_run_exports` for `qt-main`, the `poppler-qt` package somehow
correctly inherits a run dependency on it while just `poppler` doesn't,
even though both packages have it as a host dep. Wha????