recipe/build.sh: try to fix osx-arm64 build
I'm guessing a bit here since I can't reproduce the CI failure on my
system. CMake is complaining about a too-old version of libtiff, while
the versions installed in both the host and build environments are
sufficiently new. Maybe it's detecting libtiff in the wrong place? So
let's try to encourage it not to do so.